refactor: createModel.tsx
