+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ice/store",
-  "version": "1.0.0-beta.2",
+  "version": "1.0.0",
   "description": "Lightweight React state management library based on react hooks",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
