+++ src/createStore.tsx
@@ -1,201 +1,12 @@
-import React, { useState, useMemo, useEffect, useRef, useCallback } from 'react';
-import isPromise from 'is-promise';
-import transform from 'lodash.transform';
-import { createContainer } from './createContainer';
-import { PropType, ReactSetState, ContextHookReturn } from './types';
-
-const isDev = process.env.NODE_ENV !== 'production';
-
-export type ModelConfigAction<S = any> = (prevState: S, payload?: any, actions?: any, globalActions?: any) => S | Promise<S>;
-
-export interface ModelConfigActions<S = any> {
-  [name: string]: ModelConfigAction<S>;
-}
-
-export interface ModelConfig<S = any> {
-  state: S;
-  actions?: ModelConfigActions<S>;
-};
-
-export interface ModelConfigs {
-  [namespace: string]: ModelConfig;
-}
-
-export interface ModelProps<S = any> {
-  initialState?: S;
-}
-
-export interface FunctionState {
-  isLoading: boolean;
-  error: Error;
-}
-
-export type FunctionsState<Functions> = {
-  [K in keyof Functions]: FunctionState;
-}
-
-export type SetFunctionsState<Functions> = ReactSetState<FunctionsState<Functions>>;
-
-export type ActionIdentifier = number;
-
-export type ActionsIdentifier<ConfigActions> = {
-  [K in keyof ConfigActions]: ActionIdentifier;
-}
-
-export interface ActionPayload {
-  payload: any;
-  identifier: ActionIdentifier;
-}
-
-export type ActionsPayload<ConfigActions> = {
-  [K in keyof ConfigActions]: ActionPayload;
-}
-
-export type SetActionsPayload<ConfigActions> = ReactSetState<ActionsPayload<ConfigActions>>;
-
-export type Actions<ConfigActions extends ModelConfigActions> = {
-  [K in keyof ConfigActions]: (payload?: Parameters<ConfigActions[K]>[1]) => void;
-}
-
-export type TModelConfigState<M extends ModelConfig> = PropType<M, 'state'>;
-export type TModelConfigActions<M extends ModelConfig> = PropType<M, 'actions'>;
-export type TModelActions<M extends ModelConfig> = Actions<TModelConfigActions<M>>;
-export type TModelActionsState<M extends ModelConfig> = FunctionsState<TModelConfigActions<M>>;
-export type TUseModelValue<M extends ModelConfig> = [ TModelConfigState<M>, TModelActions<M>, TModelActionsState<M> ];
-export type TModel<M extends ModelConfig> =
-  ContextHookReturn<TModelConfigState<M>, TUseModelValue<M>, [(model: TUseModelValue<M>) => TModelConfigState<M>, (model: TUseModelValue<M>) => TModelActions<M>, (model: TUseModelValue<M>) => TModelActionsState<M>]>;
-
-export function createModel<M extends ModelConfig, K = string>(config: M, namespace?: K, modelsActions?): TModel<M> {
-  type ModelState = TModelConfigState<M>;
-  type ModelConfigActions = TModelConfigActions<M>;
-  type ModelConfigActionsKey = keyof ModelConfigActions;
-  type ModelActions = TModelActions<M>;
-  type ModelActionsState = TModelActionsState<M>;
-  type SetModelFunctionsState = SetFunctionsState<ModelConfigActions>;
-  type UseModelValue = TUseModelValue<M>;
-
-  const { state: defineState = {}, actions: defineActions = [] } = config;
-  let actions;
-
-  function useFunctionsState(functions: ModelConfigActionsKey[]):
-  [ ModelActionsState, SetModelFunctionsState, (name: ModelConfigActionsKey, args: FunctionState) => void ] {
-    const functionsInitialState = useMemo<ModelActionsState>(
-      () => transform(functions, (result, name) => {
-        result[name] = {
-          isLoading: false,
-          error: null,
-        };
-      }, {}),
-      [functions],
-    );
-    const [ functionsState, setFunctionsState ] = useState<ModelActionsState>(() => functionsInitialState);
-    const setFunctionState = useCallback(
-      (name: ModelConfigActionsKey, args: FunctionState) => setFunctionsState(prevState => ({
-        ...prevState,
-        [name]: {
-          ...prevState[name],
-          ...args,
-        },
-      })),
-      [],
-    );
-    return [ functionsState, setFunctionsState, setFunctionState ];
-  }
-
-  function useActions(state: ModelState, setState: ReactSetState<ModelState>):
-  [ ActionsPayload<ModelActions>, (name: ModelConfigActionsKey, payload: any) => void, ModelActionsState ] {
-    const [ actionsState, , setActionsState ] = useFunctionsState(Object.keys(defineActions));
-    const [ actionsInitialPayload, actionsInitialIdentifier ]: [ActionsPayload<ModelActions>, ActionsIdentifier<ModelActions>] = useMemo(
-      () => transform(defineActions, (result, action, name) => {
-        const state = {
-          payload: null,
-          identifier: 0,
-        };
-        result[0][name] = state;
-        result[1][name] = state.identifier;
-      }, [ {}, {} ]),
-      [],
-    );
-    const [ actionsPayload, setActionsPayload ]: [ ActionsPayload<ModelActions>, SetActionsPayload<ModelActions> ] = useState(() => actionsInitialPayload);
-    const setActionPayload = useCallback(
-      (name, payload) => setActionsPayload(prevState => ({
-        ...prevState,
-        [name]: {
-          ...prevState[name],
-          payload,
-          identifier: prevState[name].identifier + 1,
-        },
-      })),
-      [],
-    );
-
-    const actionsIdentifier = useRef(actionsInitialIdentifier);
-    const actionsPayloadIdentifier = Object.keys(actionsPayload).map((name) => actionsPayload[name].identifier);
-
-    useEffect(() => {
-      Object.keys(actionsPayload).forEach((name) => {
-        const { identifier, payload } = actionsPayload[name];
-        if (identifier && identifier !== actionsIdentifier.current[name]) {
-          actionsIdentifier.current = {
-            ...actionsIdentifier.current,
-            [name]: identifier,
-          };
-          (async () => {
-            const nextState = defineActions[name](state, payload, actions, modelsActions);
-            if (isPromise(nextState)) {
-              setActionsState(name, {
-                isLoading: true,
-                error: null,
-              });
-              try {
-                setState(await nextState);
-                setActionsState(name, {
-                  isLoading: false,
-                  error: null,
-                });
-              } catch (error) {
-                setActionsState(name, {
-                  isLoading: false,
-                  error,
-                });
-              }
-            } else {
-              setState(nextState);
-            }
-          })();
-        }
-      });
-    }, [ actionsPayloadIdentifier ]);
-
-    return [ actionsPayload, setActionPayload, actionsState ];
-  }
-
-  function useModel({ initialState }: ModelProps<ModelState>): UseModelValue {
-    const preloadedState = initialState || (defineState as ModelState);
-    const [ state, setState ] = useState<ModelState>(preloadedState);
-    const [ , executeAction, actionsState ] = useActions(state, setState);
-
-    actions = useMemo(() => transform(defineActions, (result, fn, name) => {
-      result[name] = (payload) => executeAction(name, payload);
-    }), [defineActions]);
-
-    if (namespace && modelsActions) {
-      modelsActions[namespace] = actions;
-    }
-    return [ state, actions, actionsState ];
-  }
-
-  if (isDev && namespace) {
-    useModel.displayName = namespace;
-  }
-
-  return createContainer(
-    useModel,
-  value => value[0],
-  value => value[1],
-  value => value[2],
-  );
-}
+import React from 'react';
+import { createModel } from './createModel';
+import {
+  ModelConfigs,
+  TModel,
+  TModelConfigState,
+  TModelActions,
+  TModelActionsState,
+} from './types';
 
 export function createStore<C extends ModelConfigs>(configs: C) {
   function getModel<K extends keyof C>(namespace: K): TModel<C[K]> {
