+++ src/index.ts
@@ -1,2 +1,4 @@
 export * from './createContainer';
 export * from './createStore';
+export * from './createModel';
+export * from './types';
